# STM internals overhaul