--- a/core/shared/src/main/scala/zio/stm/TArray.scala
+++ b/core/shared/src/main/scala/zio/stm/TArray.scala
@@ -81,7 +81,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Find the first element in the array matching a predicate.
    */
   def find(p: A => Boolean): USTM[Option[A]] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       var i   = 0
       var res = Option.empty[A]
 
@@ -95,13 +95,13 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
       }
 
       TExit.Succeed(res)
-    })
+    }
 
   /**
    * Find the last element in the array matching a predicate.
    */
   def findLast(p: A => Boolean): USTM[Option[A]] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       var i   = array.length - 1
       var res = Option.empty[A]
 
@@ -115,7 +115,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
       }
 
       TExit.Succeed(res)
-    })
+    }
 
   /**
    * Find the last element in the array matching a transactional predicate.
@@ -157,7 +157,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically folds using a pure function.
    */
   def fold[Z](zero: Z)(op: (Z, A) => Z): USTM[Z] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       var res = zero
       var i   = 0
 
@@ -168,7 +168,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
       }
 
       TExit.Succeed(res)
-    })
+    }
 
   /**
    * Atomically folds using a transactional function.
@@ -220,7 +220,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
     if (from < 0)
       STM.succeedNow(-1)
     else
-      ZSTM.Effect((journal, _, _) => {
+      ZSTM.Effect { (journal, _, _) =>
         var i     = from
         var found = false
 
@@ -231,7 +231,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
         }
 
         if (found) TExit.Succeed(i - 1) else TExit.Succeed(-1)
-      })
+      }
 
   /**
    * Get the index of the first entry in the array matching a transactional
@@ -267,7 +267,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
     if (end >= array.length)
       STM.succeedNow(-1)
     else
-      ZSTM.Effect((journal, _, _) => {
+      ZSTM.Effect { (journal, _, _) =>
         var i     = end
         var found = false
 
@@ -277,7 +277,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
         }
 
         if (found) TExit.Succeed(i + 1) else TExit.Succeed(-1)
-      })
+      }
 
   /**
    * The last entry in the array, if it exists.
@@ -301,7 +301,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically reduce the array, if non-empty, by a binary operator.
    */
   def reduceOption(op: (A, A) => A): USTM[Option[A]] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       var i   = 0
       var res = null.asInstanceOf[A]
 
@@ -317,7 +317,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
       }
 
       TExit.Succeed(Option(res))
-    })
+    }
 
   /**
    * Atomically reduce the non-empty array using a transactional binary operator.
@@ -352,7 +352,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically updates all elements using a pure function.
    */
   def transform(f: A => A): USTM[Unit] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       var i = 0
 
       while (i < array.length) {
@@ -362,14 +362,14 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
       }
 
       TExit.Succeed(())
-    })
+    }
 
   /**
    * Atomically updates all elements using a transactional effect.
    */
   def transformM[E](f: A => STM[E, A]): STM[E, Unit] =
     STM.foreach(Chunk.fromArray(array))(_.get.flatMap(f)).flatMap { newData =>
-      ZSTM.Effect((journal, _, _) => {
+      ZSTM.Effect { (journal, _, _) =>
         var i  = 0
         val it = newData.iterator
 
@@ -379,7 +379,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
         }
 
         TExit.Succeed(())
-      })
+      }
     }
 
   /**