--- a/core/shared/src/main/scala/zio/stm/TArray.scala
+++ b/core/shared/src/main/scala/zio/stm/TArray.scala
@@ -81,7 +81,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Find the first element in the array matching a predicate.
    */
   def find(p: A => Boolean): USTM[Option[A]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       var i   = 0
       var res = Option.empty[A]
 
@@ -101,7 +101,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Find the last element in the array matching a predicate.
    */
   def findLast(p: A => Boolean): USTM[Option[A]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       var i   = array.length - 1
       var res = Option.empty[A]
 
@@ -157,7 +157,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically folds using a pure function.
    */
   def fold[Z](zero: Z)(op: (Z, A) => Z): USTM[Z] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       var res = zero
       var i   = 0
 
@@ -220,7 +220,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
     if (from < 0)
       STM.succeedNow(-1)
     else
-      new ZSTM((journal, _, _, _) => {
+      ZSTM.Effect((journal, _, _, _) => {
         var i     = from
         var found = false
 
@@ -267,7 +267,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
     if (end >= array.length)
       STM.succeedNow(-1)
     else
-      new ZSTM((journal, _, _, _) => {
+      ZSTM.Effect((journal, _, _, _) => {
         var i     = end
         var found = false
 
@@ -301,7 +301,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically reduce the array, if non-empty, by a binary operator.
    */
   def reduceOption(op: (A, A) => A): USTM[Option[A]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       var i   = 0
       var res = null.asInstanceOf[A]
 
@@ -352,7 +352,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    * Atomically updates all elements using a pure function.
    */
   def transform(f: A => A): USTM[Unit] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       var i = 0
 
       while (i < array.length) {
@@ -369,7 +369,7 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
    */
   def transformM[E](f: A => STM[E, A]): STM[E, Unit] =
     STM.foreach(Chunk.fromArray(array))(_.get.flatMap(f)).flatMap { newData =>
-      new ZSTM((journal, _, _, _) => {
+      ZSTM.Effect((journal, _, _, _) => {
         var i  = 0
         val it = newData.iterator