--- a/core/shared/src/main/scala/zio/stm/TMap.scala
+++ b/core/shared/src/main/scala/zio/stm/TMap.scala
@@ -44,7 +44,7 @@ final class TMap[K, V] private (
    * Removes binding for given key.
    */
   def delete(k: K): USTM[Unit] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val buckets = tBuckets.unsafeGet(journal)
       val idx     = TMap.indexOf(k, buckets.array.length)
       val bucket  = buckets.array(idx).unsafeGet(journal)
@@ -64,7 +64,7 @@ final class TMap[K, V] private (
    * Atomically folds using a pure function.
    */
   def fold[A](zero: A)(op: (A, (K, V)) => A): USTM[A] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val buckets = tBuckets.unsafeGet(journal)
       var res     = zero
       var i       = 0
@@ -97,7 +97,7 @@ final class TMap[K, V] private (
    * Retrieves value associated with given key.
    */
   def get(k: K): USTM[Option[V]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val buckets = tBuckets.unsafeGet(journal)
       val idx     = TMap.indexOf(k, buckets.array.length)
       val bucket  = buckets.array(idx).unsafeGet(journal)
@@ -167,7 +167,7 @@ final class TMap[K, V] private (
       tBuckets.unsafeSet(journal, new TArray(newArray))
     }
 
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val buckets      = tBuckets.unsafeGet(journal)
       val capacity     = buckets.array.length
       val idx          = TMap.indexOf(k, capacity)
@@ -203,7 +203,7 @@ final class TMap[K, V] private (
    * Removes bindings matching predicate.
    */
   def removeIf(p: (K, V) => Boolean): USTM[Unit] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val f        = p.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -237,7 +237,7 @@ final class TMap[K, V] private (
    * Retains bindings matching predicate.
    */
   def retainIf(p: (K, V) => Boolean): USTM[Unit] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val f        = p.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -283,7 +283,7 @@ final class TMap[K, V] private (
    * Collects all bindings into a chunk.
    */
   def toChunk: USTM[Chunk[(K, V)]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
       val size     = tSize.unsafeGet(journal)
@@ -317,7 +317,7 @@ final class TMap[K, V] private (
    * Atomically updates all bindings using a pure function.
    */
   def transform(f: (K, V) => (K, V)): USTM[Unit] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val g        = f.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -365,7 +365,7 @@ final class TMap[K, V] private (
       val g = f.tupled
 
       STM.foreach(data)(g).flatMap { newData =>
-        new ZSTM((journal, _, _, _) => {
+        ZSTM.Effect((journal, _, _, _) => {
           val buckets    = tBuckets.unsafeGet(journal)
           val capacity   = buckets.array.length
           val newBuckets = Array.fill[List[(K, V)]](capacity)(Nil)