--- a/core/shared/src/main/scala/zio/stm/TMap.scala
+++ b/core/shared/src/main/scala/zio/stm/TMap.scala
@@ -44,7 +44,7 @@ final class TMap[K, V] private (
    * Removes binding for given key.
    */
   def delete(k: K): USTM[Unit] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val buckets = tBuckets.unsafeGet(journal)
       val idx     = TMap.indexOf(k, buckets.array.length)
       val bucket  = buckets.array(idx).unsafeGet(journal)
@@ -58,13 +58,13 @@ final class TMap[K, V] private (
       }
 
       TExit.unit
-    })
+    }
 
   /**
    * Atomically folds using a pure function.
    */
   def fold[A](zero: A)(op: (A, (K, V)) => A): USTM[A] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val buckets = tBuckets.unsafeGet(journal)
       var res     = zero
       var i       = 0
@@ -79,7 +79,7 @@ final class TMap[K, V] private (
       }
 
       TExit.Succeed(res)
-    })
+    }
 
   /**
    * Atomically folds using a transactional function.
@@ -97,13 +97,13 @@ final class TMap[K, V] private (
    * Retrieves value associated with given key.
    */
   def get(k: K): USTM[Option[V]] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val buckets = tBuckets.unsafeGet(journal)
       val idx     = TMap.indexOf(k, buckets.array.length)
       val bucket  = buckets.array(idx).unsafeGet(journal)
 
       TExit.Succeed(bucket.find(_._1 == k).map(_._2))
-    })
+    }
 
   /**
    * Retrieves value associated with given key or default value, in case the
@@ -167,7 +167,7 @@ final class TMap[K, V] private (
       tBuckets.unsafeSet(journal, new TArray(newArray))
     }
 
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val buckets      = tBuckets.unsafeGet(journal)
       val capacity     = buckets.array.length
       val idx          = TMap.indexOf(k, capacity)
@@ -190,7 +190,7 @@ final class TMap[K, V] private (
       }
 
       TExit.unit
-    })
+    }
   }
 
   /**
@@ -203,7 +203,7 @@ final class TMap[K, V] private (
    * Removes bindings matching predicate.
    */
   def removeIf(p: (K, V) => Boolean): USTM[Unit] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val f        = p.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -231,13 +231,13 @@ final class TMap[K, V] private (
       tSize.unsafeSet(journal, newSize)
 
       TExit.unit
-    })
+    }
 
   /**
    * Retains bindings matching predicate.
    */
   def retainIf(p: (K, V) => Boolean): USTM[Unit] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val f        = p.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -265,7 +265,7 @@ final class TMap[K, V] private (
       tSize.unsafeSet(journal, newSize)
 
       TExit.unit
-    })
+    }
 
   /**
    * Returns the number of bindings.
@@ -283,7 +283,7 @@ final class TMap[K, V] private (
    * Collects all bindings into a chunk.
    */
   def toChunk: USTM[Chunk[(K, V)]] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
       val size     = tSize.unsafeGet(journal)
@@ -305,7 +305,7 @@ final class TMap[K, V] private (
       }
 
       TExit.Succeed(Chunk.fromArray(data))
-    })
+    }
 
   /**
    * Collects all bindings into a map.
@@ -317,7 +317,7 @@ final class TMap[K, V] private (
    * Atomically updates all bindings using a pure function.
    */
   def transform(f: (K, V) => (K, V)): USTM[Unit] =
-    ZSTM.Effect((journal, _, _) => {
+    ZSTM.Effect { (journal, _, _) =>
       val g        = f.tupled
       val buckets  = tBuckets.unsafeGet(journal)
       val capacity = buckets.array.length
@@ -355,7 +355,7 @@ final class TMap[K, V] private (
       tSize.unsafeSet(journal, newSize)
 
       TExit.unit
-    })
+    }
 
   /**
    * Atomically updates all bindings using a transactional function.
@@ -365,7 +365,7 @@ final class TMap[K, V] private (
       val g = f.tupled
 
       STM.foreach(data)(g).flatMap { newData =>
-        ZSTM.Effect((journal, _, _) => {
+        ZSTM.Effect { (journal, _, _) =>
           val buckets    = tBuckets.unsafeGet(journal)
           val capacity   = buckets.array.length
           val newBuckets = Array.fill[List[(K, V)]](capacity)(Nil)
@@ -391,7 +391,7 @@ final class TMap[K, V] private (
 
           tSize.unsafeSet(journal, newSize)
           TExit.unit
-        })
+        }
       }
     }