--- a/core/shared/src/main/scala/zio/stm/TPriorityQueue.scala
+++ b/core/shared/src/main/scala/zio/stm/TPriorityQueue.scala
@@ -47,7 +47,7 @@ final class TPriorityQueue[A] private (private val ref: TRef[SortedMap[A, ::[A]]
    * a value is in the queue.
    */
   def peek: USTM[A] =
-    new ZSTM((journal, _, _, _) =>
+    ZSTM.Effect((journal, _, _, _) =>
       ref.unsafeGet(journal).headOption match {
         case None          => TExit.Retry
         case Some((_, as)) => TExit.Succeed(as.head)
@@ -90,7 +90,7 @@ final class TPriorityQueue[A] private (private val ref: TRef[SortedMap[A, ::[A]]
    * Takes a value from the queue, retrying until a value is in the queue.
    */
   def take: USTM[A] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val map = ref.unsafeGet(journal)
       map.headOption match {
         case None => TExit.Retry
@@ -143,7 +143,7 @@ final class TPriorityQueue[A] private (private val ref: TRef[SortedMap[A, ::[A]]
    * the queue.
    */
   def takeOption: USTM[Option[A]] =
-    new ZSTM((journal, _, _, _) => {
+    ZSTM.Effect((journal, _, _, _) => {
       val map = ref.unsafeGet(journal)
       map.headOption match {
         case None => TExit.Succeed(None)