… core/shared/src/main/scala/zio/stm/TQueue.scala
… core/shared/src/main/scala/zio/stm/TReentrantLock.scala
… core/shared/src/main/scala/zio/stm/ZSTM.scala
… core/shared/src/main/scala/zio/stm/ZTRef.scala
… project/MimaSettings.scala