This pull request changes the encoding of ZSTM to a declarative one, similar to the one used in ZIO, intending to improve its performance.

The following sections contain benchmark results. The latest version from master - 590c24c0f9ceaf14c337367fbf8d2fb39542212d, is used as baseline.

### TArray

#### Before

```
Benchmark                          (size)   Mode  Cnt       Score        Error  Units
TArrayOpsBenchmarks.find               10  thrpt   15  179549.856 ±   2254.001  ops/s
TArrayOpsBenchmarks.find              100  thrpt   15  100904.365 ±  36388.462  ops/s
TArrayOpsBenchmarks.find             1000  thrpt   15    3776.653 ±     35.453  ops/s
TArrayOpsBenchmarks.find            10000  thrpt   15     238.890 ±      3.947  ops/s
TArrayOpsBenchmarks.find           100000  thrpt   15      23.592 ±      4.625  ops/s
TArrayOpsBenchmarks.findM              10  thrpt   15   94526.589 ±   1375.946  ops/s
TArrayOpsBenchmarks.findM             100  thrpt   15   11963.031 ±    511.167  ops/s
TArrayOpsBenchmarks.findM            1000  thrpt   15    3010.418 ±      8.302  ops/s
TArrayOpsBenchmarks.findM           10000  thrpt   15      91.556 ±      3.066  ops/s
TArrayOpsBenchmarks.findM          100000  thrpt   15       3.150 ±      0.053  ops/s
TArrayOpsBenchmarks.fold               10  thrpt   15  478163.249 ±    407.561  ops/s
TArrayOpsBenchmarks.fold              100  thrpt   15   46480.346 ±    692.277  ops/s
TArrayOpsBenchmarks.fold             1000  thrpt   15    3525.044 ±     37.371  ops/s
TArrayOpsBenchmarks.fold            10000  thrpt   15     520.609 ±    207.883  ops/s
TArrayOpsBenchmarks.fold           100000  thrpt   15       6.033 ±      0.298  ops/s
TArrayOpsBenchmarks.foldM              10  thrpt   15   91136.232 ±    924.846  ops/s
TArrayOpsBenchmarks.foldM             100  thrpt   15   15872.687 ±   8841.261  ops/s
TArrayOpsBenchmarks.foldM            1000  thrpt   15     817.120 ±    414.170  ops/s
TArrayOpsBenchmarks.foldM           10000  thrpt   15      38.311 ±      1.422  ops/s
TArrayOpsBenchmarks.foldM          100000  thrpt   15       2.894 ±      1.578  ops/s
TArrayOpsBenchmarks.indexWhere         10  thrpt   15  282897.341 ± 146569.141  ops/s
TArrayOpsBenchmarks.indexWhere        100  thrpt   15   47112.732 ±    554.623  ops/s
TArrayOpsBenchmarks.indexWhere       1000  thrpt   15    3581.373 ±     47.226  ops/s
TArrayOpsBenchmarks.indexWhere      10000  thrpt   15     524.009 ±    205.139  ops/s
TArrayOpsBenchmarks.indexWhere     100000  thrpt   15       6.425 ±      0.234  ops/s
TArrayOpsBenchmarks.indexWhereM        10  thrpt   15  128710.757 ±   1720.708  ops/s
TArrayOpsBenchmarks.indexWhereM       100  thrpt   15   56104.418 ±  13148.243  ops/s
TArrayOpsBenchmarks.indexWhereM      1000  thrpt   15    1460.726 ±     28.721  ops/s
TArrayOpsBenchmarks.indexWhereM     10000  thrpt   15     111.934 ±      4.932  ops/s
TArrayOpsBenchmarks.indexWhereM    100000  thrpt   15      10.479 ±      0.290  ops/s
TArrayOpsBenchmarks.lookup             10  thrpt   15    2898.011 ±     39.818  ops/s
TArrayOpsBenchmarks.lookup            100  thrpt   15    2899.231 ±     34.567  ops/s
TArrayOpsBenchmarks.lookup           1000  thrpt   15    7742.991 ±     17.094  ops/s
TArrayOpsBenchmarks.lookup          10000  thrpt   15    2152.589 ±     31.201  ops/s
TArrayOpsBenchmarks.lookup         100000  thrpt   15    2822.675 ±     23.449  ops/s
TArrayOpsBenchmarks.reduceOption       10  thrpt   15  410328.927 ± 135850.986  ops/s
TArrayOpsBenchmarks.reduceOption      100  thrpt   15   45997.991 ±    875.609  ops/s
TArrayOpsBenchmarks.reduceOption     1000  thrpt   15    3169.543 ±     32.357  ops/s
TArrayOpsBenchmarks.reduceOption    10000  thrpt   15     396.042 ±    215.278  ops/s
TArrayOpsBenchmarks.reduceOption   100000  thrpt   15       7.680 ±      4.748  ops/s
TArrayOpsBenchmarks.reduceOptionM      10  thrpt   15   81335.803 ±   1007.464  ops/s
TArrayOpsBenchmarks.reduceOptionM     100  thrpt   15    8976.763 ±    405.668  ops/s
TArrayOpsBenchmarks.reduceOptionM    1000  thrpt   15    1087.285 ±    516.888  ops/s
TArrayOpsBenchmarks.reduceOptionM   10000  thrpt   15      37.012 ±      1.442  ops/s
TArrayOpsBenchmarks.reduceOptionM  100000  thrpt   15       2.223 ±      0.048  ops/s
TArrayOpsBenchmarks.transform          10  thrpt   15  322845.421 ± 134588.203  ops/s
TArrayOpsBenchmarks.transform         100  thrpt   15   34240.216 ±    355.018  ops/s
TArrayOpsBenchmarks.transform        1000  thrpt   15    2451.817 ±     20.586  ops/s
TArrayOpsBenchmarks.transform       10000  thrpt   15     438.771 ±     55.657  ops/s
TArrayOpsBenchmarks.transform      100000  thrpt   15       3.904 ±      0.130  ops/s
TArrayOpsBenchmarks.transformM         10  thrpt   15   78933.497 ±    858.017  ops/s
TArrayOpsBenchmarks.transformM        100  thrpt   15   25735.062 ±    214.456  ops/s
TArrayOpsBenchmarks.transformM       1000  thrpt   15     751.115 ±     12.811  ops/s
TArrayOpsBenchmarks.transformM      10000  thrpt   15      58.295 ±      1.522  ops/s
TArrayOpsBenchmarks.transformM     100000  thrpt   15       6.224 ±      0.195  ops/s
TArrayOpsBenchmarks.update             10  thrpt   15    2723.976 ±     25.886  ops/s
TArrayOpsBenchmarks.update            100  thrpt   15    2706.516 ±     19.234  ops/s
TArrayOpsBenchmarks.update           1000  thrpt   15    6601.722 ±   1623.339  ops/s
TArrayOpsBenchmarks.update          10000  thrpt   15    2531.806 ±     20.980  ops/s
TArrayOpsBenchmarks.update         100000  thrpt   15    2560.499 ±     21.071  ops/s
TArrayOpsBenchmarks.updateM            10  thrpt   15    3252.176 ±   1572.536  ops/s
TArrayOpsBenchmarks.updateM           100  thrpt   15    1632.590 ±     19.241  ops/s
TArrayOpsBenchmarks.updateM          1000  thrpt   15    1741.424 ±     18.927  ops/s
TArrayOpsBenchmarks.updateM         10000  thrpt   15    2035.552 ±   1036.124  ops/s
TArrayOpsBenchmarks.updateM        100000  thrpt   15    2494.702 ±   1313.301  ops/s
```

#### After

```
Benchmark                          (size)   Mode  Cnt       Score       Error  Units
TArrayOpsBenchmarks.find               10  thrpt   15  479496.180 ±   615.206  ops/s
TArrayOpsBenchmarks.find              100  thrpt   15  124181.092 ±   150.847  ops/s
TArrayOpsBenchmarks.find             1000  thrpt   15    9857.831 ±    28.309  ops/s
TArrayOpsBenchmarks.find            10000  thrpt   15     656.120 ±     3.308  ops/s
TArrayOpsBenchmarks.find           100000  thrpt   15      24.073 ±     1.174  ops/s
TArrayOpsBenchmarks.findM              10  thrpt   15  303843.483 ±   377.320  ops/s
TArrayOpsBenchmarks.findM             100  thrpt   15   51237.792 ±    99.112  ops/s
TArrayOpsBenchmarks.findM            1000  thrpt   15    3967.423 ±     6.453  ops/s
TArrayOpsBenchmarks.findM           10000  thrpt   15     348.757 ±     3.058  ops/s
TArrayOpsBenchmarks.findM          100000  thrpt   15      12.566 ±     0.673  ops/s
TArrayOpsBenchmarks.fold               10  thrpt   15  477514.963 ±   593.632  ops/s
TArrayOpsBenchmarks.fold              100  thrpt   15  121218.496 ±   184.899  ops/s
TArrayOpsBenchmarks.fold             1000  thrpt   15    9703.641 ±    31.260  ops/s
TArrayOpsBenchmarks.fold            10000  thrpt   15     645.616 ±     5.220  ops/s
TArrayOpsBenchmarks.fold           100000  thrpt   15      22.138 ±     1.600  ops/s
TArrayOpsBenchmarks.foldM              10  thrpt   15  265368.275 ±   706.619  ops/s
TArrayOpsBenchmarks.foldM             100  thrpt   15   42409.511 ±    60.470  ops/s
TArrayOpsBenchmarks.foldM            1000  thrpt   15    3444.914 ±     8.487  ops/s
TArrayOpsBenchmarks.foldM           10000  thrpt   15     325.850 ±     1.557  ops/s
TArrayOpsBenchmarks.foldM          100000  thrpt   15      10.549 ±     0.254  ops/s
TArrayOpsBenchmarks.indexWhere         10  thrpt   15  475493.516 ±   857.737  ops/s
TArrayOpsBenchmarks.indexWhere        100  thrpt   15  126688.685 ±    93.029  ops/s
TArrayOpsBenchmarks.indexWhere       1000  thrpt   15   10011.308 ±    43.110  ops/s
TArrayOpsBenchmarks.indexWhere      10000  thrpt   15     673.131 ±     5.321  ops/s
TArrayOpsBenchmarks.indexWhere     100000  thrpt   15      22.505 ±     2.022  ops/s
TArrayOpsBenchmarks.indexWhereM        10  thrpt   15  377975.967 ±   353.525  ops/s
TArrayOpsBenchmarks.indexWhereM       100  thrpt   15   73549.969 ±   115.304  ops/s
TArrayOpsBenchmarks.indexWhereM      1000  thrpt   15    5298.496 ±     7.691  ops/s
TArrayOpsBenchmarks.indexWhereM     10000  thrpt   15     459.579 ±    11.128  ops/s
TArrayOpsBenchmarks.indexWhereM    100000  thrpt   15      12.882 ±     0.379  ops/s
TArrayOpsBenchmarks.lookup             10  thrpt   15    7352.442 ±    51.625  ops/s
TArrayOpsBenchmarks.lookup            100  thrpt   15    7491.974 ±    10.985  ops/s
TArrayOpsBenchmarks.lookup           1000  thrpt   15    7470.940 ±     7.828  ops/s
TArrayOpsBenchmarks.lookup          10000  thrpt   15    7262.242 ±     8.449  ops/s
TArrayOpsBenchmarks.lookup         100000  thrpt   15    5777.528 ±    20.192  ops/s
TArrayOpsBenchmarks.reduceOption       10  thrpt   15  473261.824 ±  3670.340  ops/s
TArrayOpsBenchmarks.reduceOption      100  thrpt   15  120809.769 ±   117.493  ops/s
TArrayOpsBenchmarks.reduceOption     1000  thrpt   15    9564.268 ±    27.791  ops/s
TArrayOpsBenchmarks.reduceOption    10000  thrpt   15     656.598 ±     3.985  ops/s
TArrayOpsBenchmarks.reduceOption   100000  thrpt   15      22.356 ±     1.177  ops/s
TArrayOpsBenchmarks.reduceOptionM      10  thrpt   15  221073.968 ±  1132.185  ops/s
TArrayOpsBenchmarks.reduceOptionM     100  thrpt   15   33960.814 ±    40.824  ops/s
TArrayOpsBenchmarks.reduceOptionM    1000  thrpt   15    2939.036 ±     6.406  ops/s
TArrayOpsBenchmarks.reduceOptionM   10000  thrpt   15     265.613 ±     1.284  ops/s
TArrayOpsBenchmarks.reduceOptionM  100000  thrpt   15      10.741 ±     0.358  ops/s
TArrayOpsBenchmarks.transform          10  thrpt   15  414375.899 ±   344.412  ops/s
TArrayOpsBenchmarks.transform         100  thrpt   15   33895.537 ±   332.888  ops/s
TArrayOpsBenchmarks.transform        1000  thrpt   15    5615.220 ±  2085.861  ops/s
TArrayOpsBenchmarks.transform       10000  thrpt   15     178.360 ±     2.509  ops/s
TArrayOpsBenchmarks.transform      100000  thrpt   15       4.146 ±     0.134  ops/s
TArrayOpsBenchmarks.transformM         10  thrpt   15  158099.269 ± 82096.536  ops/s
TArrayOpsBenchmarks.transformM        100  thrpt   15   13277.839 ±  4559.803  ops/s
TArrayOpsBenchmarks.transformM       1000  thrpt   15    1110.351 ±    22.878  ops/s
TArrayOpsBenchmarks.transformM      10000  thrpt   15     100.314 ±    46.059  ops/s
TArrayOpsBenchmarks.transformM     100000  thrpt   15       3.567 ±     1.987  ops/s
TArrayOpsBenchmarks.update             10  thrpt   15    2660.324 ±    49.534  ops/s
TArrayOpsBenchmarks.update            100  thrpt   15    2668.692 ±    37.798  ops/s
TArrayOpsBenchmarks.update           1000  thrpt   15    5387.682 ±  2169.463  ops/s
TArrayOpsBenchmarks.update          10000  thrpt   15    2641.401 ±    27.150  ops/s
TArrayOpsBenchmarks.update         100000  thrpt   15    2431.362 ±    29.276  ops/s
TArrayOpsBenchmarks.updateM            10  thrpt   15    4346.987 ±   794.557  ops/s
TArrayOpsBenchmarks.updateM           100  thrpt   15    1776.712 ±    36.313  ops/s
TArrayOpsBenchmarks.updateM          1000  thrpt   15    1666.502 ±    19.049  ops/s
TArrayOpsBenchmarks.updateM         10000  thrpt   15    4333.688 ±     2.518  ops/s
TArrayOpsBenchmarks.updateM        100000  thrpt   15    1582.372 ±    12.367  ops/s
```

### TReentrantLock

#### Before

```
Benchmark                                                              Mode  Cnt        Score         Error  Units
TReentrantLockBenchmark.JavaLockBasic                                 thrpt   20   668079.294 ±    8563.079  ops/s
TReentrantLockBenchmark.JavaLockBasic:javaLockReadGroup               thrpt   20    48466.025 ±    6825.909  ops/s
TReentrantLockBenchmark.JavaLockBasic:javaLockWriteGroup              thrpt   20   619613.269 ±    6481.720  ops/s
TReentrantLockBenchmark.JavaLockHighContention                        thrpt   20  2543805.382 ± 1495997.146  ops/s
TReentrantLockBenchmark.JavaLockHighContention:javaLockReadGroup3     thrpt   20  1577175.953 ±  998734.499  ops/s
TReentrantLockBenchmark.JavaLockHighContention:javaLockWriteGroup3    thrpt   20   966629.429 ±  523226.595  ops/s
TReentrantLockBenchmark.JavaLockLowContention                         thrpt   20  1720706.097 ±  898311.791  ops/s
TReentrantLockBenchmark.JavaLockLowContention:javaLockReadGroup1      thrpt   20  1420807.187 ±  752032.037  ops/s
TReentrantLockBenchmark.JavaLockLowContention:javaLockWriteGroup1     thrpt   20   299898.910 ±  156663.955  ops/s
TReentrantLockBenchmark.JavaLockMediumContention                      thrpt   20  2371901.528 ± 1527448.086  ops/s
TReentrantLockBenchmark.JavaLockMediumContention:javaLockReadGroup2   thrpt   20  1712245.750 ± 1120002.331  ops/s
TReentrantLockBenchmark.JavaLockMediumContention:javaLockWriteGroup2  thrpt   20   659655.777 ±  412053.850  ops/s
TReentrantLockBenchmark.ZioLockBasic                                  thrpt   20    80724.832 ±   60831.806  ops/s
TReentrantLockBenchmark.ZioLockBasic:zioLockReadGroup                 thrpt   20    40427.531 ±   30476.993  ops/s
TReentrantLockBenchmark.ZioLockBasic:zioLockWriteGroup                thrpt   20    40297.301 ±   30354.957  ops/s
TReentrantLockBenchmark.ZioLockHighContention                         thrpt   20    43918.724 ±    1555.958  ops/s
TReentrantLockBenchmark.ZioLockHighContention:zioLockReadGroup3       thrpt   20    21886.606 ±     868.890  ops/s
TReentrantLockBenchmark.ZioLockHighContention:zioLockWriteGroup3      thrpt   20    22032.119 ±     891.499  ops/s
TReentrantLockBenchmark.ZioLockLowContention                          thrpt   20    80485.318 ±   60173.781  ops/s
TReentrantLockBenchmark.ZioLockLowContention:zioLockReadGroup1        thrpt   20    64304.512 ±   48103.124  ops/s
TReentrantLockBenchmark.ZioLockLowContention:zioLockWriteGroup1       thrpt   20    16180.806 ±   12078.240  ops/s
TReentrantLockBenchmark.ZioLockMediumContention                       thrpt   20    43722.252 ±    1982.662  ops/s
TReentrantLockBenchmark.ZioLockMediumContention:zioLockReadGroup2     thrpt   20    29248.573 ±    1393.106  ops/s
TReentrantLockBenchmark.ZioLockMediumContention:zioLockWriteGroup2    thrpt   20    14473.680 ±     854.821  ops/s
```

#### After

```
Benchmark                                                              Mode  Cnt        Score         Error  Units
TReentrantLockBenchmark.JavaLockBasic                                 thrpt   20  1014183.922 ±  707157.803  ops/s
TReentrantLockBenchmark.JavaLockBasic:javaLockReadGroup               thrpt   20    51750.570 ±    8973.856  ops/s
TReentrantLockBenchmark.JavaLockBasic:javaLockWriteGroup              thrpt   20   962433.353 ±  701791.603  ops/s
TReentrantLockBenchmark.JavaLockHighContention                        thrpt   20  2479245.548 ± 1432173.674  ops/s
TReentrantLockBenchmark.JavaLockHighContention:javaLockReadGroup3     thrpt   20  1448461.354 ±  882274.533  ops/s
TReentrantLockBenchmark.JavaLockHighContention:javaLockWriteGroup3    thrpt   20  1030784.194 ±  588351.655  ops/s
TReentrantLockBenchmark.JavaLockLowContention                         thrpt   20  1427840.896 ±   52057.474  ops/s
TReentrantLockBenchmark.JavaLockLowContention:javaLockReadGroup1      thrpt   20  1138802.500 ±   83655.569  ops/s
TReentrantLockBenchmark.JavaLockLowContention:javaLockWriteGroup1     thrpt   20   289038.397 ±   34764.968  ops/s
TReentrantLockBenchmark.JavaLockMediumContention                      thrpt   20  2703598.607 ± 1609031.234  ops/s
TReentrantLockBenchmark.JavaLockMediumContention:javaLockReadGroup2   thrpt   20  1903896.382 ± 1150230.796  ops/s
TReentrantLockBenchmark.JavaLockMediumContention:javaLockWriteGroup2  thrpt   20   799702.225 ±  462236.764  ops/s
TReentrantLockBenchmark.ZioLockBasic                                  thrpt   20    61323.795 ±   43116.090  ops/s
TReentrantLockBenchmark.ZioLockBasic:zioLockReadGroup                 thrpt   20    30502.052 ±   21484.410  ops/s
TReentrantLockBenchmark.ZioLockBasic:zioLockWriteGroup                thrpt   20    30821.742 ±   21632.236  ops/s
TReentrantLockBenchmark.ZioLockHighContention                         thrpt   20    74381.605 ±   57956.523  ops/s
TReentrantLockBenchmark.ZioLockHighContention:zioLockReadGroup3       thrpt   20    37415.115 ±   28950.419  ops/s
TReentrantLockBenchmark.ZioLockHighContention:zioLockWriteGroup3      thrpt   20    36966.490 ±   29019.289  ops/s
TReentrantLockBenchmark.ZioLockLowContention                          thrpt   20    74221.771 ±   60545.223  ops/s
TReentrantLockBenchmark.ZioLockLowContention:zioLockReadGroup1        thrpt   20    59258.412 ±   48443.790  ops/s
TReentrantLockBenchmark.ZioLockLowContention:zioLockWriteGroup1       thrpt   20    14963.359 ±   12113.298  ops/s
TReentrantLockBenchmark.ZioLockMediumContention                       thrpt   20    69202.718 ±   56326.866  ops/s
TReentrantLockBenchmark.ZioLockMediumContention:zioLockReadGroup2     thrpt   20    46234.532 ±   37646.704  ops/s
TReentrantLockBenchmark.ZioLockMediumContention:zioLockWriteGroup2    thrpt   20    22968.186 ±   18688.454  ops/s
```

### TMap vs TRef contention

#### Before

```
Benchmark                               (repeatedUpdates)   Mode  Cnt    Score   Error  Units
TMapContentionBenchmarks.contentionMap                100  thrpt   15  141.378 ± 0.533  ops/s
TMapContentionBenchmarks.contentionMap               1000  thrpt   15   14.696 ± 0.080  ops/s
TMapContentionBenchmarks.contentionMap              10000  thrpt   15    1.514 ± 0.016  ops/s
TMapContentionBenchmarks.contentionRef                100  thrpt   15  155.414 ± 0.831  ops/s
TMapContentionBenchmarks.contentionRef               1000  thrpt   15   16.956 ± 0.052  ops/s
TMapContentionBenchmarks.contentionRef              10000  thrpt   15    1.705 ± 0.007  ops/s
```

#### After

```
Benchmark                               (repeatedUpdates)   Mode  Cnt    Score   Error  Units
TMapContentionBenchmarks.contentionMap                100  thrpt   15  139.635 ± 0.450  ops/s
TMapContentionBenchmarks.contentionMap               1000  thrpt   15   14.376 ± 0.039  ops/s
TMapContentionBenchmarks.contentionMap              10000  thrpt   15    1.469 ± 0.011  ops/s
TMapContentionBenchmarks.contentionRef                100  thrpt   15  152.830 ± 0.535  ops/s
TMapContentionBenchmarks.contentionRef               1000  thrpt   15    9.040 ± 5.913  ops/s
TMapContentionBenchmarks.contentionRef              10000  thrpt   15    0.532 ± 0.092  ops/s
```

### TMap

#### Before

```
Benchmark                     (size)   Mode  Cnt       Score        Error  Units
TMapOpsBenchmarks.fold             0  thrpt   15  266643.523 ± 135525.766  ops/s
TMapOpsBenchmarks.fold            10  thrpt   15  260146.778 ± 125049.286  ops/s
TMapOpsBenchmarks.fold           100  thrpt   15   66289.916 ±  33957.553  ops/s
TMapOpsBenchmarks.fold          1000  thrpt   15    5517.232 ±   2884.312  ops/s
TMapOpsBenchmarks.fold         10000  thrpt   15     212.703 ±    124.214  ops/s
TMapOpsBenchmarks.fold        100000  thrpt   15       7.675 ±      5.359  ops/s
TMapOpsBenchmarks.foldM            0  thrpt   15  195224.560 ± 107673.762  ops/s
TMapOpsBenchmarks.foldM           10  thrpt   15  152904.102 ±  80814.611  ops/s
TMapOpsBenchmarks.foldM          100  thrpt   15   28840.842 ±  15308.443  ops/s
TMapOpsBenchmarks.foldM         1000  thrpt   15    1701.679 ±    846.569  ops/s
…