@robmwalsh 2020-10-31
@mijicd just took a quick look at a random selection of some of the numbers... Niiiice! This morning I have some tracing work planned then I've got friends visiting for the next few days. Won't be able to review until Wednesday sorry, but I'm looking forward to it!