@robmwalsh 2020-10-31
Speaking of tracing, if the encoding is similar to ZIO...