@mijicd 2020-10-31
Two remarks:

- I had to disable mima checks for STM. Otherwise, this change is a no go.
- Dotty "disagrees" with some of the changes. I have to investigate, but if someone spots the issue, please shout.