@robmwalsh 2020-10-31
@mijicd I don't think we can just break bincompat on STM since it's part of core, but there may be ways around the mima issues. See https://github.com/zio/zio/pull/4352#issuecomment-719692030 as one example that may help