@mijicd 2020-10-31
**Update:**

Dotty issues are resolved by adding some typecasts where it was necessary. Check out 4213101dfcdc7215bd827549a555d1f8caabc75b for details.

Binary compatibility issues are unfortunately still there, and thus the checks are ignored. To understand the problem, consider the current ZSTM encoding:

```scala
final class ZSTM[-R, +E, +A] private[stm] (private val exec (ZSTM.internal.Journal, Fiber.Id, AtomicLong, R) => ZSTM.internal.TExit[E, A]) extends AnyVal { ... }
```

In this encoding, ZSTM is a `Function4`. Changing the encoding to sealed trait effectively changed the return type of all functions present in stm package.