@robmwalsh 2020-11-01
well the obvious solution to that particular problem is to make it a Function4!

```scala
sealed trait ZSTM[-R, +E, +A]
    extends Serializable
    with ((ZSTM.internal.Journal, Fiber.Id, AtomicLong, R) => ZSTM.internal.TExit[E, A]) { ??? }
```
That takes the mima issues down to 3, which are just that your cases don't have a suitable apply method. If you can add an apply method that makes sense for each of them, you're golden :D
```
[IJ]mimaChecks
[info] compiling 2 Scala sources to /home/rob/IdeaProjects/zio/core/jvm/target/scala-2.12/classes ...
[error] /home/rob/IdeaProjects/zio/core/shared/src/main/scala/zio/stm/ZSTM.scala:1460:33: class FoldCauseM needs to be abstract, since method apply in trait Function4 of type (v1: zio.stm.ZSTM.internal.Journal, v2: zio.Fiber.Id, v3: java.util.concurrent.atomic.AtomicLong, v4: R)zio.stm.ZSTM.internal.TExit[E2,B] is not defined
[error]   private[stm] final case class FoldCauseM[R, E1, E2, A, B](
[error]                                 ^
[error] /home/rob/IdeaProjects/zio/core/shared/src/main/scala/zio/stm/ZSTM.scala:1470:33: class Effect needs to be abstract, since method apply in trait Function4 of type (v1: zio.stm.ZSTM.internal.Journal, v2: zio.Fiber.Id, v3: java.util.concurrent.atomic.AtomicLong, v4: R)zio.stm.ZSTM.internal.TExit[E,A] is not defined
[error]   private[stm] final case class Effect[R, E, A](f: (Journal, Fiber.Id, R) => A) extends ZSTM[R, E, A]
[error]                                 ^
[error] /home/rob/IdeaProjects/zio/core/shared/src/main/scala/zio/stm/ZSTM.scala:1472:33: class ProvideSome needs to be abstract, since method apply in trait Function4 of type (v1: zio.stm.ZSTM.internal.Journal, v2: zio.Fiber.Id, v3: java.util.concurrent.atomic.AtomicLong, v4: R2)zio.stm.ZSTM.internal.TExit[E,A] is not defined
[error]   private[stm] final case class ProvideSome[R1, R2, E, A](effect: ZSTM[R1, E, A], f: R2 => R1) extends ZSTM[R2, E, A]
[error]
```